code-revision: moving switch statement cases to modular implementation pathspec
As titled.
- README.md updated accordingly.
