+++ implementation/index.js
@@ -1,0 +1,9 @@
+import Ring from './ring';
+import Wireframe from './wireframe';
+import RightTriangle from './right-triangle';
+
+export {
+    Ring,
+    Wireframe,
+    RightTriangle,
+}
